// src/types.ts
export interface ClassType {
  id: string;
  name: string;
  batchSize: number; 
}